Move inventory and quest definitions into their own files. Add a conclusion to the bat quest.
